+++ src/app/code/community/Zip/Payment/Helper/Data.php
@@ -150,7 +150,7 @@ class Zip_Payment_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function isOneStepCheckout()
     {
-        return Mage::helper('zip_payment')->getConfig()->getValue(Zip_Payment_Model_Config::CONFIG_CHECKOUT_TYPE_PATH) == Zip_Payment_Model_Adminhtml_System_Config_Source_CheckoutType::CHECK_TYPE_ONE_STEP;
+        return Mage::helper('zip_payment')->getConfig()->getValue(Zip_Payment_Model_Config::CONFIG_CHECKOUT_TYPE_PATH) == Zip_Payment_Model_Adminhtml_System_Config_Source_CheckoutType::CHECKOUT_TYPE_ONE_STEP;
     }
 
     /**