+++ src/app/code/community/Zip/Payment/Model/Config.php
@@ -70,6 +70,8 @@ class Zip_Payment_Model_Config
     const CONFIG_CHECKOUT_TYPE_PATH = 'payment/zip_payment/checkout/type';
     const CONFIG_CHECKOUT_GENERAL_ERROR_PATH = 'payment/zip_payment/checkout/error/general';
     const CONFIG_CHECKOUT_JS_LIB_PATH = 'payment/zip_payment/checkout/js_lib';
+    const CONFIG_CHECKOUT_STYLE_PATH = 'payment/zip_payment/checkout/style';
+
     const CONFIG_CHECKOUT_DISPLAY_MODE_PATH = 'payment/zip_payment/checkout/display_mode';
     const CONFIG_CHECKOUT_CUSTOM_SCRIPT_PATH = 'payment/zip_payment/checkout/custom_script';